% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\docType{methods}
\name{Run}
\alias{Run}
\alias{Run,SlurmRunner-method}
\title{Run the R cmd as script and keep track of the process}
\usage{
Run(x, filename, cmd, local = FALSE, ...)
}
\arguments{
\item{x}{the SlurmRunner object}

\item{filename}{the filename to save the R script to (has to be unique for the analysis!)}

\item{cmd}{the R cmd to run (use absolute paths in there!)}

\item{local}{true to run using bash (sbatch default)}
}
\value{
the downstream PID of the spawned R script
}
\description{
run the script using either a SLUEM backend or the local computer
}
