% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.R
\docType{methods}
\name{running}
\alias{running}
\alias{running,SlurmRunner-method}
\title{fetch the still running pid's}
\usage{
running(x)
}
\arguments{
\item{x}{the SlurmRunner object}
}
\description{
Check which of the internally stored pids are still active
}
