% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script.R
\docType{methods}
\name{Script}
\alias{Script}
\alias{Script,SlurmRunner-method}
\title{Create the script files to run the R cmd}
\usage{
Script(x, filename, cmd, ...)
}
\arguments{
\item{filename}{the filename to save the R script to (has to be unique for the analysis!)}

\item{cmd}{The R cmd to run}
}
\value{
the SLURM/batch script to run the R code
}
\description{
use the cmd as script workload and built the SLURM header for the script using the internal settings hash
}
