% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlurmRunner_class.R
\docType{methods}
\name{SlurmRunner}
\alias{SlurmRunner}
\alias{SlurmRunner,SlurmRunner-method}
\title{description of function SlurmRunner}
\usage{
SlurmRunner(settings = list(), tmp.path = "", debug = FALSE)
}
\arguments{
\item{settings}{the settings for the SLURM process (only 'A', 't' and 'p' are used)}

\item{tmp.path}{where to store the temporaray files}
}
\value{
A new SlurmRunner object
}
\description{
create a new SlurmRunner object. The clustering will be performed on the columns of the data.frame.
}
