% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlurmRunner_class.R
\docType{package}
\name{SlurmRunner}
\alias{SlurmRunner}
\alias{SlurmRunner-package}
\title{SlurmRunner}
\description{
An S4 class to run SLURM scripts and keep track of the running ones
}
\section{Slots}{

\describe{
\item{\code{running_local}}{the local still running scripts}

\item{\code{running_slurm}}{the slurm running scripts}

\item{\code{settings}}{the slurm settings to be used in the scripts}

\item{\code{tmp.path}}{the run folder where all slurm scripts are placed}

\item{\code{debug}}{if true the scripts are not run (default FALSE)}

\item{\code{uname}}{the username of the active user (default whoami output)}
}}

